Rename external to ext
ext is shrtr.
