+++ antenna/ext/fs/crashstorage.py
@@ -9,7 +9,7 @@ import os.path
 
 from everett.component import ConfigOptions
 
-from antenna.external.crashstorage_base import CrashStorageBase
+from antenna.ext.crashstorage_base import CrashStorageBase
 from antenna.util import get_date_from_crash_id, json_ordered_dumps
