+++ antenna/ext/s3/connection.py
