+++ prod.env
@@ -11,7 +11,7 @@ STATSD_HOST=statsd
 STATSD_NAMESPACE=mcboatface
 
 # BreakdpadSubmitterResource settings
-CRASHSTORAGE_CLASS=antenna.external.s3.crashstorage.S3CrashStorage
+CRASHSTORAGE_CLASS=antenna.ext.s3.crashstorage.S3CrashStorage
 
 # S3CrashStorage and S3Connection settings
 CRASHSTORAGE_ENDPOINT_URL=http://fakes3:4569
